Adding pytest fixtures to conftest
